+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/CustomHeightDataGrid.test.jsx
@@ -88,7 +88,6 @@ class CustomHeightDataGrid extends React.Component {
     };
   }
 
-  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="custom-height-data-grid" className={cx('content-wrapper')}>
@@ -107,7 +106,6 @@ class CustomHeightDataGrid extends React.Component {
       </div>
     );
   }
-  /* eslint-enable react/forbid-dom-props */
 }
 
 export default CustomHeightDataGrid;