+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/NoFillDataGrid.test.jsx
@@ -87,7 +87,6 @@ class NoFillDataGrid extends React.Component {
     };
   }
 
-  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="no-fill-data-grid" className={cx(['content-wrapper', 'content-overflow'])}>
@@ -103,7 +102,6 @@ class NoFillDataGrid extends React.Component {
       </div>
     );
   }
-  /* eslint-enable react/forbid-dom-props */
 }
 
 export default NoFillDataGrid;