+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/NoOverflowColumnDataGrid.test.jsx
@@ -44,7 +44,6 @@ class NoOverflowColumnDataGrid extends React.Component {
     };
   }
 
-  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="no-overflow-column-data-grid" className={cx('content-wrapper')}>
@@ -60,7 +59,6 @@ class NoOverflowColumnDataGrid extends React.Component {
       </div>
     );
   }
-  /* eslint-enable react/forbid-dom-props */
 }
 
 export default NoOverflowColumnDataGrid;