+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/NoOverflowDataGrid.test.jsx
@@ -43,7 +43,6 @@ class NoOverflowDataGrid extends React.Component {
     };
   }
 
-  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="no-overflow-data-grid" className={cx('content-wrapper')}>
@@ -60,7 +59,6 @@ class NoOverflowDataGrid extends React.Component {
       </div>
     );
   }
-  /* eslint-enable react/forbid-dom-props */
 }
 
 export default NoOverflowDataGrid;