+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/PagedContentDataGrid.test.jsx
@@ -86,7 +86,6 @@ class PagedContentDataGrid extends React.Component {
     };
   }
 
-  /* eslint-disable react/forbid-dom-props */
   render() {
     return (
       <div id="paged-data-grid" className={cx('content-wrapper')}>
@@ -103,7 +102,6 @@ class PagedContentDataGrid extends React.Component {
       </div>
     );
   }
-  /* eslint-enable react/forbid-dom-props */
 }
 
 export default PagedContentDataGrid;