+++ packages/terra-clinical-data-grid/src/terra-dev-site/test/clinical-data-grid/SelectableDataGrid.test.jsx
@@ -93,7 +93,6 @@ class SelectableDataGrid extends React.Component {
     };
   }
 
-  /* eslint-disable react/forbid-dom-props */
   render() {
     const { columns } = this.state;
 
@@ -162,7 +161,6 @@ class SelectableDataGrid extends React.Component {
       </div>
     );
   }
-  /* eslint-enable react/forbid-dom-props */
 }
 
 export default SelectableDataGrid;