+++ packages/terra-clinical-item-view/src/terra-dev-site/test/clinical-item-view/OverflowDisplaysItemView.test.jsx
@@ -19,7 +19,6 @@ const accessoryStart = id => <IconAlert id={id} />;
 const accessoryEnd = id => <IconInformation id={id} />;
 const comment = id => <ItemView.Comment id={id} text="comment1comment1comment1comment1comment1comment1comment1comment1comment1comment1comment1comment1" />;
 
-/* eslint-disable react/forbid-dom-props */
 const views = () => (
   <div className={cx('overflow-displays-itemview')}>
     <p>Applied width of 900px to show the default vs truncated styling.</p>
@@ -35,6 +34,5 @@ const views = () => (
     <ItemView displays={displays} isTruncated comment={comment(4)} startAccessory={accessoryStart(8)} endAccessory={accessoryEnd(12)} layout="twoColumns" textEmphasis="start" id="ItemView-two-truncate" />
   </div>
 );
-/* eslint-enable react/forbid-dom-props */
 
 export default views;