+++ packages/terra-clinical-label-value-view/CHANGELOG.md
@@ -4,13 +4,13 @@ ChangeLog
 Unreleased
 ----------
 ### Changed
+* Removed inline styles from dev site components.
 * Add comments for ESlint 6
 * Cleaned up imports in examples and test files
 
 3.7.0 - (August 14, 2019)
 ----------
 ### Changed
-* Removed inline styles from dev site components.
 * Updated package.json test scripts
 
 3.6.0 - (July 16, 2019)