+++ packages/terra-clinical-label-value-view/CHANGELOG.md
@@ -7,7 +7,7 @@ Unreleased
 * Removed inline styles from dev site components.
 * Add comments for ESlint 6
 * Cleaned up imports in examples and test files
-* Removed clinical-item-display from examples.
+* Removed clinical-item-display as a dependency.
 
 3.7.0 - (August 14, 2019)
 ----------